requirements.txt
openai-whisper>=20240930
faster-whisper>=1.0.0
ffmpeg-python==0.2.0
transformers>=4.44.0
sentencepiece==0.2.0
protobuf>=4.25.0
torch>=2.1.0
pyannote.audio>=4.0.0
python-dotenv==1.0.0
